requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
build==0.10.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
circus==0.18.0
click==8.1.3
click-option-group==0.5.5
cloudpickle==2.2.1
contextlib2==21.6.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
deepmerge==1.1.0
Deprecated==1.2.13
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
facenet-pytorch==2.5.3
flasgger==0.9.5
Flask==2.2.3
flask-babel==3.1.0
Flask-Cors==3.0.10
Flask-Executor==1.0.0
Flask-WTF==1.1.1
fonttools==4.39.4
frozenlist==1.3.3
fs==2.4.16
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mistune==2.0.5
mongoengine==0.27.0
multidict==6.0.4
numpy==1.24.3
oauth2client==4.1.3
opencv-python==4.7.0.72
opentelemetry-api==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
pip-requirements-parser==32.0.1
pip-tools==6.13.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.16.0
protobuf==4.23.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyDrive==1.3.1
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
pynvml==11.5.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
rich==13.3.5
rsa==4.9
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
simple-di==0.1.5
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.114
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
